Move EquipType subclasses to modules/Equipment
- Move definitions of all EquipType subclasses into modules/Equipment/Types.lua
- EquipType.lua now returns the EquipType class directly.
- Add ShieldType subclass for shield generators.
