+++ data/modules/Equipment/Utility.lua
@@ -1,7 +1,7 @@
 -- Copyright © 2008-2025 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local EquipTypes = require 'EquipType'
+local EquipTypes = require '.Types'
 local Equipment = require 'Equipment'
 
 local EquipType = EquipTypes.EquipType
